requirements.txt
streamlit>=1.28.0
requests>=2.31.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0